Testing individual files for each sensor
